move detailed pv to examples
