+++ examples/Detailed_PV_Layout/detailed_pv_config.py
