+++ examples/Wind_Floris/simulate_hybrid_wbattery_dispatch_custom.py
